requirements-light.txt
# Core dependencies
beautifulsoup4==4.12.3
lxml==5.1.0
requests==2.31.0
httpx==0.26.0  # Async HTTP client for parallel scraping
playwright==1.41.0

# NLP & Text Processing (lightweight)
flashtext==2.7
scikit-learn==1.4.0

# Data Processing
pandas==2.2.0
numpy==1.26.3
pydantic==2.6.0

# Utilities
joblib==1.3.2
pyyaml==6.0.1
python-dotenv==1.0.0
diskcache==5.6.3

# Google Sheets
gspread==6.0.0
google-auth==2.27.0

# Scraping helpers
feedparser==6.0.10
